move syscall name into utils
